perf(core): read a block's edits from the index, not the whole log
ensure_doc and materialize_text_from_log each walked every op in the log
to rebuild one block's text, so the first keystroke on a cold block after
a full-replay boot paid an O(217k) scan. Both now route through one
doc_from_log built on OpLog::edit_updates, which removes the duplication
and both scans together.

This is the third instance of the bug #179 fixed in block_text, so it
comes with a guard. A scan and an index return identical bytes — which is
exactly why it came back twice and why no correctness test can catch it —
so the guard measures the ratio across log sizes rather than a duration,
and does not care how fast the machine is.
